requirements.txt
fastapi
uvicorn
pydantic-ai
httpx
requests
python-dotenv